+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- (Internal) Add new repository function to update security properties.
+
 ### Changed
 
 - (Internal) Refactor Transaction into new repository pattern.